ポートレートモードを試験的機能から外観設定へ移設して本番でも利用可能にする
