+++ src/store/atoms/display.ts
@@ -1,5 +1,5 @@
 import { atom } from 'jotai';
 
-// 試験的機能のオンオフ。フィールド単位のプリミティブatomとして公開し、
+// 画面表示に関する設定。フィールド単位のプリミティブatomとして公開し、
 // 読み取りは必ずこちらを購読する(docs/state-management.md 参照)。
 export const portraitModeEnabledAtom = atom(false);
